requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chart-studio==1.1.0
click==7.1.2
cryptography==3.4.7
dash==1.19.0
dash-auth==1.4.1
dash-bootstrap-components==0.11.2
dash-core-components==1.15.0
dash-extensions==0.0.51
dash-html-components==1.1.2
dash-leaflet==0.1.13
dash-renderer==1.9.0
dash-table==4.11.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.1
django-allauth==0.44.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap-pagination==1.7.1
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
filelock==3.0.12
Flask==1.1.2
Flask-Caching==1.10.1
Flask-Compress==1.8.0
Flask-SeaSurf==0.3.0
future==0.18.2
geobuf==1.1.1
geographiclib==1.52
geopy==2.2.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
more-itertools==8.7.0
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.2
pbr==5.5.1
Pillow==7.2.0
plotly==4.14.3
protobuf==3.14.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
Shapely==1.7.1
six==1.15.0
soupsieve==2.2.1
sqlparse==0.3.1
stevedore==3.3.0
ua-parser==0.10.0
urllib3==1.26.3
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
Werkzeug==1.0.1